Add option for natural sorting and make it default
The search(...) function creates per default a natural sorted array.
This behaviour is now enforced by adding an explicit 'natural'
parameter. Sorting by the ascii table in now an option in the config
menu.

Add files for the configuration and add a test for the 'natural'
sorting.

However the ascii sorting is still broken and will be fixed in a later
commit. Currently the pages of an open namespace are not always below
the open namespace but below another one.