Fix position of subpages in ascii sorting
Since colons come in the ascii table after numbers, there was some
unintended sorting with multiple namespaces, which only differed by
unindented numbers at the end. Replacing the colons with the '/' which
is in front of the numbers in the ascii table solved the problem.

Add a test for this.

Known possible issue: pages are not sorted after namespaces.